+++ storage/store/partial_kv.go
@@ -8,11 +8,9 @@ import (
 	"math/big"
 
 	"github.com/shopspring/decimal"
-	pbssinternal "github.com/streamingfast/substreams/pb/sf/substreams/intern/v2"
 	pbsubstreams "github.com/streamingfast/substreams/pb/sf/substreams/v1"
 	"github.com/streamingfast/substreams/storage/store/marshaller"
 	"go.uber.org/zap"
-	"google.golang.org/protobuf/proto"
 )
 
 var _ Store = (*PartialKV)(nil)
@@ -20,7 +18,6 @@ var _ Store = (*PartialKV)(nil)
 type PartialKV struct {
 	*baseStore
 
-	operations      *pbssinternal.Operations
 	initialBlock    uint64 // block at which we initialized this store
 	DeletedPrefixes []string
 
@@ -86,12 +83,6 @@ func (p *PartialKV) Save(endBoundaryBlock uint64) (*FileInfo, *fileWriter, error
 }
 
 func (p *PartialKV) DeletePrefix(ord uint64, prefix string) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type: pbssinternal.Operation_DELETE_PREFIX,
-		Ord:  ord,
-		Key:  prefix,
-	})
-
 	p.baseStore.DeletePrefix(ord, prefix)
 
 	if !p.seen[prefix] {
@@ -111,23 +102,6 @@ func (p *PartialKV) DeleteStore(ctx context.Context, file *FileInfo) (err error)
 
 func (p *PartialKV) String() string {
 	return fmt.Sprintf("partialKV name %s moduleInitialBlock %d  keyCount %d deltasCount %d loadFrom %s", p.Name(), p.moduleInitialBlock, len(p.kv), len(p.deltas), p.loadedFrom)
-}
-
-func (p *PartialKV) Reset() {
-	p.operations = &pbssinternal.Operations{}
-	p.baseStore.Reset()
-}
-
-func (p *PartialKV) ApplyOps(in []byte) error {
-	return applyOps(in, p.baseStore)
-}
-
-func (p *PartialKV) ReadOps() []byte {
-	data, err := proto.Marshal(p.operations)
-	if err != nil {
-		panic(err)
-	}
-	return data
 }
 
 func valueToFloat64(value []byte) float64 {
@@ -155,286 +129,10 @@ func valueToBigDecimal(value []byte) (decimal.Decimal, error) {
 	return val, nil
 }
 
-func applyOps(in []byte, store *baseStore) error {
-	ops := &pbssinternal.Operations{}
-	if err := proto.Unmarshal(in, ops); err != nil {
-		return err
-	}
-
-	for _, op := range ops.Operations {
-		switch op.Type {
-		case pbssinternal.Operation_SET:
-			store.Set(op.Ord, op.Key, string(op.Value))
-		case pbssinternal.Operation_SET_BYTES:
-			store.SetBytes(op.Ord, op.Key, op.Value)
-		case pbssinternal.Operation_SET_IF_NOT_EXISTS:
-			store.SetIfNotExists(op.Ord, op.Key, string(op.Value))
-		case pbssinternal.Operation_SET_BYTES_IF_NOT_EXISTS:
-			store.SetBytesIfNotExists(op.Ord, op.Key, op.Value)
-		case pbssinternal.Operation_APPEND:
-			store.Append(op.Ord, op.Key, op.Value)
-		case pbssinternal.Operation_DELETE_PREFIX:
-			store.DeletePrefix(op.Ord, op.Key)
-		case pbssinternal.Operation_SET_MAX_BIG_INT:
-			store.SetMaxBigInt(op.Ord, op.Key, valueToBigInt(op.Value))
-		case pbssinternal.Operation_SET_MAX_INT64:
-			store.SetMaxInt64(op.Ord, op.Key, valueToInt64(op.Value))
-		case pbssinternal.Operation_SET_MAX_FLOAT64:
-			store.SetMaxFloat64(op.Ord, op.Key, valueToFloat64(op.Value))
-		case pbssinternal.Operation_SET_MAX_BIG_DECIMAL:
-			val, err := valueToBigDecimal(op.Value)
-			if err != nil {
-				return err
-			}
-			store.SetMaxBigDecimal(op.Ord, op.Key, val)
-		case pbssinternal.Operation_SET_MIN_BIG_INT:
-			store.SetMinBigInt(op.Ord, op.Key, valueToBigInt(op.Value))
-		case pbssinternal.Operation_SET_MIN_INT64:
-			store.SetMinInt64(op.Ord, op.Key, valueToInt64(op.Value))
-		case pbssinternal.Operation_SET_MIN_FLOAT64:
-			store.SetMinFloat64(op.Ord, op.Key, valueToFloat64(op.Value))
-		case pbssinternal.Operation_SET_MIN_BIG_DECIMAL:
-			val, err := valueToBigDecimal(op.Value)
-			if err != nil {
-				return err
-			}
-			store.SetMinBigDecimal(op.Ord, op.Key, val)
-		case pbssinternal.Operation_SUM_BIG_INT:
-			store.SumBigInt(op.Ord, op.Key, valueToBigInt(op.Value))
-		case pbssinternal.Operation_SUM_INT64:
-			store.SumInt64(op.Ord, op.Key, valueToInt64(op.Value))
-		case pbssinternal.Operation_SUM_FLOAT64:
-			store.SumFloat64(op.Ord, op.Key, valueToFloat64(op.Value))
-		case pbssinternal.Operation_SUM_BIG_DECIMAL:
-			val, err := valueToBigDecimal(op.Value)
-			if err != nil {
-				return err
-			}
-			store.SumBigDecimal(op.Ord, op.Key, val)
-		}
-	}
-	return nil
-}
-
 func (p *PartialKV) ApplyDelta(delta *pbsubstreams.StoreDelta) {
 	panic("caching store cannot be used with deltas")
 }
 
 func (p *PartialKV) ApplyDeltasReverse(deltas []*pbsubstreams.StoreDelta) {
 	panic("caching store cannot be used with deltas")
-}
-
-// apparently this is faster than append() method
-func cloneBytes(b []byte) []byte {
-	out := make([]byte, len(b))
-	copy(out, b)
-	return out
-}
-
-func bigIntToBytes(i *big.Int) []byte {
-	return []byte(i.String())
-}
-
-func float64ToBytes(f float64) []byte {
-	var buf [8]byte
-	binary.LittleEndian.PutUint64(buf[:], math.Float64bits(f))
-	return buf[:]
-}
-
-func int64ToBytes(i int64) []byte {
-	big := new(big.Int)
-	big.SetInt64(i)
-	return []byte(big.String())
-}
-
-func bigDecimalToBytes(d decimal.Decimal) []byte {
-	val, err := d.MarshalBinary()
-	if err != nil {
-		panic(err)
-	}
-	return val
-}
-
-func (p *PartialKV) Set(ord uint64, key string, value string) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET,
-		Ord:   ord,
-		Key:   key,
-		Value: cloneBytes([]byte(value)),
-	})
-
-	p.baseStore.Set(ord, key, value)
-}
-
-func (p *PartialKV) SetBytes(ord uint64, key string, value []byte) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_BYTES,
-		Ord:   ord,
-		Key:   key,
-		Value: cloneBytes(value),
-	})
-
-	p.baseStore.SetBytes(ord, key, value)
-}
-
-func (p *PartialKV) SetIfNotExists(ord uint64, key string, value string) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_IF_NOT_EXISTS,
-		Ord:   ord,
-		Key:   key,
-		Value: cloneBytes([]byte(value)),
-	})
-
-	p.baseStore.SetIfNotExists(ord, key, value)
-}
-
-func (p *PartialKV) SetBytesIfNotExists(ord uint64, key string, value []byte) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_BYTES_IF_NOT_EXISTS,
-		Ord:   ord,
-		Key:   key,
-		Value: cloneBytes(value),
-	})
-
-	p.baseStore.SetBytesIfNotExists(ord, key, value)
-}
-
-func (p *PartialKV) Append(ord uint64, key string, value []byte) error {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_APPEND,
-		Ord:   ord,
-		Key:   key,
-		Value: cloneBytes(value),
-	})
-
-	return p.baseStore.Append(ord, key, value)
-}
-
-func (p *PartialKV) SetMaxBigInt(ord uint64, key string, value *big.Int) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_MAX_BIG_INT,
-		Ord:   ord,
-		Key:   key,
-		Value: bigIntToBytes(value),
-	})
-
-	p.baseStore.SetMaxBigInt(ord, key, value)
-}
-
-func (p *PartialKV) SetMaxInt64(ord uint64, key string, value int64) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_MAX_INT64,
-		Ord:   ord,
-		Key:   key,
-		Value: int64ToBytes(value),
-	})
-	p.baseStore.SetMaxInt64(ord, key, value)
-}
-
-func (p *PartialKV) SetMaxFloat64(ord uint64, key string, value float64) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_MAX_FLOAT64,
-		Ord:   ord,
-		Key:   key,
-		Value: float64ToBytes(value),
-	})
-
-	p.baseStore.SetMaxFloat64(ord, key, value)
-}
-
-func (p *PartialKV) SetMaxBigDecimal(ord uint64, key string, value decimal.Decimal) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_MAX_BIG_DECIMAL,
-		Ord:   ord,
-		Key:   key,
-		Value: bigDecimalToBytes(value),
-	})
-
-	p.baseStore.SetMaxBigDecimal(ord, key, value)
-}
-
-func (p *PartialKV) SetMinBigInt(ord uint64, key string, value *big.Int) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_MIN_BIG_INT,
-		Ord:   ord,
-		Key:   key,
-		Value: bigIntToBytes(value),
-	})
-	p.baseStore.SetMinBigInt(ord, key, value)
-}
-
-func (p *PartialKV) SetMinInt64(ord uint64, key string, value int64) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_MIN_INT64,
-		Ord:   ord,
-		Key:   key,
-		Value: int64ToBytes(value),
-	})
-	p.baseStore.SetMinInt64(ord, key, value)
-}
-
-func (p *PartialKV) SetMinFloat64(ord uint64, key string, value float64) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_MIN_FLOAT64,
-		Ord:   ord,
-		Key:   key,
-		Value: float64ToBytes(value),
-	})
-
-	p.baseStore.SetMinFloat64(ord, key, value)
-}
-
-func (p *PartialKV) SetMinBigDecimal(ord uint64, key string, value decimal.Decimal) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SET_MIN_BIG_DECIMAL,
-		Ord:   ord,
-		Key:   key,
-		Value: bigDecimalToBytes(value),
-	})
-
-	p.baseStore.SetMinBigDecimal(ord, key, value)
-}
-
-func (p *PartialKV) SumBigInt(ord uint64, key string, value *big.Int) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SUM_BIG_INT,
-		Ord:   ord,
-		Key:   key,
-		Value: bigIntToBytes(value),
-	})
-
-	p.baseStore.SumBigInt(ord, key, value)
-}
-
-func (p *PartialKV) SumInt64(ord uint64, key string, value int64) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SUM_INT64,
-		Ord:   ord,
-		Key:   key,
-		Value: int64ToBytes(value),
-	})
-
-	p.baseStore.SumInt64(ord, key, value)
-}
-
-func (p *PartialKV) SumFloat64(ord uint64, key string, value float64) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SUM_FLOAT64,
-		Ord:   ord,
-		Key:   key,
-		Value: float64ToBytes(value),
-	})
-
-	p.baseStore.SumFloat64(ord, key, value)
-}
-
-func (p *PartialKV) SumBigDecimal(ord uint64, key string, value decimal.Decimal) {
-	p.operations.Operations = append(p.operations.Operations, &pbssinternal.Operation{
-		Type:  pbssinternal.Operation_SUM_BIG_DECIMAL,
-		Ord:   ord,
-		Key:   key,
-		Value: bigDecimalToBytes(value),
-	})
-
-	p.baseStore.SumBigDecimal(ord, key, value)
 }