+++ storage/store/store_max_test.go
@@ -10,6 +10,7 @@ import (
 	pbsubstreams "github.com/streamingfast/substreams/pb/sf/substreams/v1"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestStoreSetMaxBigInt(t *testing.T) {
@@ -59,6 +60,7 @@ func TestStoreSetMaxBigInt(t *testing.T) {
 			initTestStore(test.store, test.key, test.existingValue) // b.kv = []{}
 
 			test.store.SetMaxBigInt(0, test.key, test.value)
+			require.NoError(t, test.store.Flush())
 			actual, found := test.store.GetAt(0, test.key)
 			if !found {
 				t.Errorf("value not found")
@@ -125,6 +127,7 @@ func TestStoreSetMaxInt64(t *testing.T) {
 			initTestStore(test.store, test.key, test.existingValue)
 
 			test.store.SetMaxInt64(0, test.key, test.value)
+			require.NoError(t, test.store.Flush())
 			actual, found := test.store.GetAt(0, test.key)
 			if !found {
 				t.Errorf("value not found")
@@ -192,6 +195,7 @@ func TestStoreSetMaxFloat64(t *testing.T) {
 			initTestStore(test.store, test.key, test.existingValue)
 
 			test.store.SetMaxFloat64(0, test.key, test.value)
+			require.NoError(t, test.store.Flush())
 			actual, found := test.store.GetAt(0, test.key)
 			if !found {
 				t.Errorf("value not found")
@@ -253,6 +257,7 @@ func TestStoreSetMaxBigFloat(t *testing.T) {
 			initTestStore(test.store, test.key, test.existingValue)
 
 			test.store.SetMaxBigDecimal(0, test.key, test.value)
+			require.NoError(t, test.store.Flush())
 			actual, found := test.store.GetAt(0, test.key)
 			if !found {
 				t.Errorf("value not found")